+++ lib/active_registration.rb
@@ -2,5 +2,4 @@ require "active_registration/version"
 require "active_registration/engine"
 
 module ActiveRegistration
-  autoload :UserExtensions, "active_registration/user_extensions"
 end